Debian has fixed all current supported releases
The security tracker requires a bit of inside knowledge to read.

If something is fixed in "bullseye-security", it is fixed (DSA issued, packages available), even if bullseye is listed as vulnerable (this refers to the main archive). For any user, it's fixed.

strecth, buster, bullseye are the supported (old)stable releases. bookworm is Debian testing and does not get security support/is not a stable release yet.

Thijs Kinkhorst
SURFcert & Debian Security Team