Removed BeyondTrust from supplier list (#288)
* Removed BeyondTrust from supplier list

BeyondTrust software was never affected by this vulnerability as there is no possible attack vector to exploit it from either product. The removal of the library and environment variable patch were only proactive measures for added security. The security bulletin will be updated shortly.

* Changed status to Not Vulnerable

Changed status to Not Vulnerable

* Changed status to Not vuln

Changed status to Not vuln