Add Detection Guidance document (#478)
* Upload Detection Guidance

Developed a guidance document which explains in which stages of the attack which detections can take place and which conclusions can be derived from it.

* Update README.md

Added the Generic Detection Guidance document, powered by Deloitte. The link points to the attack detection guidance document (pdf) which is proposed to also be included onto the repo. I can also store it on my personal fork and point it to that direction if that's more desired.

* Delete Log4j Attack Detection Guidance - Release.pdf

* Updated the Detection Guidance PDF

Made some slight changes as a reference to GovCERT.ch was missing.